master/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
boto3==1.35.48
botocore==1.35.48
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.5.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
executing==2.1.0
idna==3.10
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.7
numpy==2.0.2
openpyxl==3.1.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycryptodome==3.15.0
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
XlsxWriter==3.1.6